Adds console matrix_driver and animation library
The animation library is the library for loading the confid that defines
what intros, walks, and outros are available. The console version of the
matrix_driver is a way to visualize what the crosswlak will look like
but in the console as opposed on on the box itself. To make all of this
work, I also implmeneted the idea of time-based timers that measure the
length of intros, walks, and outros based on their audio files (and adds
the requirements that ALL gifs MUST have an audio file -- I made silent
audio files for the silent gifs).

This also makes a bunch of changes to the controller for better handling
of keyboard interrupt shutdowns in the console and to the timer to get
the state changes working with the animations.